Update the remote IP when connection is lost
When one end of the link is eg. a consumer internet connection, dynamic
DNS can be used to automatically get the new IP.
It's not clear under what circumstances a fresh request is actually
issued, though, so in a fit of paranoia this patch uses dnspython to
send the request directly to the upstream nameserver.

This is invoked whenever the connection seems to be down ie. we get an
ICMP Unreachable response (hence a socket error) or we haven't received
a packet in a while (30s, while the keepalives are at 10s interval). DNS
queries are rate limited to avoid rude behaviour.